Update *_MAX constants to reflect changes to Buffer::changed_line_max

+++ src/block.c
@@ -236,12 +236,11 @@ static size_t insert_bytes(const char *buf, size_t len)
 void do_insert(const char *buf, size_t len)
 {
     size_t nl = insert_bytes(buf, len);
-
     buffer->nl += nl;
     sanity_check();
 
     view_update_cursor_y(view);
-    buffer_mark_lines_changed(view->buffer, view->cy, nl ? INT_MAX : view->cy);
+    buffer_mark_lines_changed(view->buffer, view->cy, nl ? LONG_MAX : view->cy);
     if (buffer->syn) {
         hl_insert(buffer, view->cy, nl);
     }
@@ -339,7 +338,7 @@ char *do_delete(size_t len)
     buffer_mark_lines_changed (
         view->buffer,
         view->cy,
-        deleted_nl ? INT_MAX : view->cy
+        deleted_nl ? LONG_MAX : view->cy
     );
     if (buffer->syn) {
         hl_delete(buffer, view->cy, deleted_nl);
@@ -395,7 +394,7 @@ char *do_replace(size_t del, const char *buf, size_t ins)
         // Some line(s) changed but lines after them did not move up or down
         buffer_mark_lines_changed(view->buffer, view->cy, view->cy + del_nl);
     } else {
-        buffer_mark_lines_changed(view->buffer, view->cy, INT_MAX);
+        buffer_mark_lines_changed(view->buffer, view->cy, LONG_MAX);
     }
     if (buffer->syn) {
         hl_delete(buffer, view->cy, del_nl);
